+++ scripts/lib/utilities/templateUtils.js
@@ -4,8 +4,7 @@ function getTokensInShape(shape, scene, {x: offsetX, y: offsetY}={x: 0, y: 0}) {
     if (!shape && !scene) return tokens;
     let sceneTokens = scene.tokens;
     for (let token of sceneTokens) {
-        let pointCollisions = token.getOccupiedGridSpaceOffsets()
-            .map(i => scene.grid.getCenterPoint(i))
+        let pointCollisions = getTokenPoints(token)
             .map(i => ({x: i.x - offsetX, y: i.y - offsetY}))
             .filter(i => shape.contains(i.x, i.y));
         for (let point of pointCollisions) {
@@ -24,13 +23,31 @@ function getTemplatesInToken(token) {
     let scene = token?.document?.parent;
     if (!scene) return templates;
     let sceneTemplates = scene.templates;
-    let pointsToTest = token.document.getOccupiedGridSpaceOffsets().map(i => scene.grid.getCenterPoint(i));
+    let pointsToTest = getTokenPoints(token.document);
     for (let template of sceneTemplates) {
         if (template.object?.shape && pointsToTest.some(i => template.object.testPoint(i))) {
             templates.add(template);
         }
     }
     return templates;
+}
+function getTokenPoints(token) {
+    let points = [];
+    let grid = token.parent?.grid;
+    if (!grid) return points;
+    if (!grid.isGridless) 
+        return token.getOccupiedGridSpaceOffsets().map(i => grid.getCenterPoint(i));
+    let startX = token.width >= 1 ? 0.5 : (token.width / 2);
+    let startY = token.height >= 1 ? 0.5 : (token.height / 2);
+    for (let dx = startX; dx < token.width; dx++) {
+        for (let dy = startY; dy < token.height; dy++) {
+            points.push({
+                x: token.x + dx * grid.size,
+                y: token.y + dy * grid.size
+            });
+        }
+    }
+    return points;
 }
 function findGrids(A, B, template) {
     let locations = new Set();